+++ core/pfe-core/controllers/internals-controller.ts
@@ -1,0 +1,21 @@
+import type { ReactiveController, ReactiveControllerHost } from 'lit';
+
+export class InternalsController implements ReactiveController {
+  #internals: ElementInternals;
+
+  constructor(
+    public host: ReactiveControllerHost & HTMLElement,
+  ) {
+    this.#internals = host.attachInternals();
+  }
+
+  hostConnected?(): void
+
+  submit() {
+    this.#internals.form?.requestSubmit(this.host);
+  }
+
+  reset() {
+    this.#internals.form?.reset();
+  }
+}