+++ elements/pfe-badge/demo/pfe-badge.js
@@ -1,6 +1,5 @@
 import '@patternfly/pfe-badge';
 import '@patternfly/pfe-band';
-import '@patternfly/pfe-codeblock';
 import '@patternfly/pfe-select';
 
 const root = document.querySelector('[data-demo="pfe-badge"]')?.shadowRoot ?? document;