+++ elements/pfe-clipboard/custom-elements-manifest.config.js
@@ -1,5 +1,0 @@
-import { pfeCustomElementsManifestConfig } from '@patternfly/pfe-tools/custom-elements-manifest.js';
-
-export default pfeCustomElementsManifestConfig({
-  globs: ['pfe-*.ts'],
-});